Split out RepoLabelChecker
