+++ run.rb
@@ -28,86 +28,10 @@ def abort_if_errors
   exit 1 if @errors.positive?
 end
 
-def github_api_session
-  @github_api_session ||= begin
-    http = Net::HTTP.new('api.github.com', 443)
-    http.use_ssl = true
-    http.start
-    http
-  end
-end
-
-def user_and_password(file)
-  @user_and_password ||= begin
-    creds = JSON.parse(File.read(file))
-    "#{creds.fetch('github').fetch('user')}:#{creds.fetch('github').fetch('pass')}"
-  end
-end
-
-def do_request(klass, url, expected_status, body = nil)
-  uri = URI.parse(url)
-  req = klass.new(uri)
-
-  req['Authorization'] = if (file = ENV['DEBUG_CREDENTIALS_PATH'])
-                           "Basic #{[user_and_password(file)].pack('m0')}"
-                         else
-                           "Bearer #{ENV.fetch('GITHUB_TOKEN')}"
-                         end
-
-  req['Accept'] = 'application/vnd.github.v3+json'
-
-  if body
-    req.body = JSON.generate(body)
-    req['Content-Type'] = 'application/json'
-  end
-
-  res = github_api_session.request(req)
-  return JSON.parse(res.body) if res.code == expected_status.to_s
-
-  raise <<~MESSAGE
-    #{req.method} #{url} -> HTTP/#{res.http_version} #{res.code} #{res.message} (expected #{expected_status})
-  MESSAGE
-end
-
-def get(url)
-  do_request(Net::HTTP::Get, url, 200)
-end
-
-def post(url, body)
-  do_request(Net::HTTP::Post, url, 201, body)
-end
-
-def patch(url, body)
-  do_request(Net::HTTP::Patch, url, 200, body)
-end
-
 def ensure_labels_present(strict:)
-  # https://docs.github.com/en/rest/reference/issues#list-labels-for-a-repository
-  # https://docs.github.com/en/rest/reference/issues#create-a-label
-  # https://docs.github.com/en/rest/reference/issues#update-a-label
-
-  # FIXME: pagination. We've set the max per_page but there still might be more than that.
-  labels_in_repo = get("https://api.github.com/repos/#{repo_full_name}/labels?per_page=100")
-  warn 'Warning: 100 labels found; there might be more. FIXME, pagination' if labels_in_repo.count >= 100
-
-  by_name = labels_in_repo.to_h { |label| [label.fetch('name'), label] }
-
-  RISK_LABELS.each do |name, (color, description)|
-    existing = by_name[name]
-
-    if existing.nil?
-      body = { name: name, color: color, description: description }
-      # FIXME: race condition here, if we run concurrently on the same repo
-      post("https://api.github.com/repos/#{repo_full_name}/labels", body)
-    elsif strict && (existing.fetch('color').downcase != color.downcase || existing.fetch('description') != description)
-      body = { new_name: name, color: color, description: description }
-      patch(existing.fetch('url'), body)
-    end
-  end
-end
-
-def repo_full_name
-  event.fetch('repository').fetch('full_name')
+  require_relative './github_client'
+  require_relative './repo_label_checker'
+  RepoLabelChecker.new(event, GithubClient.new).run(strict: strict, definitions: RISK_LABELS)
 end
 
 def event
